requirements.txt
OpenCV==4.10.0
NumPy==2.2.1
PyTorch==2.5.1
PyTorch Lightning==2.5.0.post0
TorchVision==0.20.1
TorchMetrics==1.6.1
Pandas==2.2.3
Matplotlib==3.10.0
Plotly==5.24.1
git+https://github.com/facebookresearch/detectron2.git@v0.6